requirements.txt
streamlit == 1.10.0
pandas == 1.4.3
openpyxl == 3.0.10
numpy == 1.23.1
matplotlib == 3.5.2
japanize-matplotlib == 1.1.3
plotly == 5.9.0
statsmodels == 0.13.2
scikit-learn == 1.1.1
Pillow == 9.2.0